requirements.txt
# LangChain e dependências
langchain>=0.3.23
langchain-core==0.3.54
langchain-community==0.3.21
langchain-openai==0.3.14
langchain-chroma==0.2.3
python-dotenv==1.0.0
sqlalchemy==2.0.40
#chromadb==0.6.3
langsmith==0.3.32
tiktoken==0.9.0
# Web framework
streamlit>=1.44.1
# Banco de dados
psycopg2-binary>=2.9.9  # Para PostgreSQL
# Processamento de dados
pandas>=2.2.3









